+++ src/main/java/com/prueba/demo/core/model/DetalleVenta.java
@@ -1,5 +1,7 @@
 package com.prueba.demo.core.model;
 
+import java.math.BigDecimal;
+
 import lombok.Getter;
 import lombok.Setter;
 
@@ -11,7 +13,7 @@ public class DetalleVenta extends ModeloBase{
     private Integer idVenta;
     private String tipoLadrillo;
     private Integer cantidadTotal;
-    private Integer precio;
+    private BigDecimal precio;
     private String estado;
     
 }
